+++ api/init.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 npx knex migrate:latest
 npx knex seed:run
-npm run start
+npm run dev