+++ api/src/database/migrations/20201007194055_create_table_categories.js
@@ -1,13 +1,12 @@
 const { onUpdateTrigger } = require('../../../knexfile');
 
-exports.up = async knex => knex.schema.createTable('articles', table => {
+exports.up = async knex => knex.schema.createTable('categories', table => {
     table.increments('id');
     table.text('title').unique().notNullable();
     table.text('slug').unique().notNullable();
-    table.text('body').notNullable();
 
     table.timestamp('created_at').defaultTo(knex.fn.now());
     table.timestamp('updated_at').defaultTo(knex.fn.now());
-}).then(() => knex.raw(onUpdateTrigger('articles')));
+}).then(() => knex.raw(onUpdateTrigger('categories')));
 
-exports.down = async knex => knex.schema.dropTable('articles');
+exports.down = async knex => knex.schema.dropTable('categories');